% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_codons_to_aa.R
\name{dna_codons_to_aa}
\alias{dna_codons_to_aa}
\title{Convert DNA codons to Amino Acids}
\usage{
dna_codons_to_aa(codons)
}
\arguments{
\item{codons}{}
}
\value{
Amino Acids
}
\description{
Convert DNA codons to Amino Acids
}
\examples{
dna_codons_to_aa(codons)
}
