% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_codons.R
\name{mk_codons}
\alias{mk_codons}
\title{Create Codons}
\usage{
mk_codons(dna, s = 1)
}
\arguments{
\item{l}{length of desired DNA string}
}
\value{
DNA string
}
\description{
This function creates codons in a DNA string
}
\examples{
mk_codons(dna, s=1)

}
