% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_dna.R
\name{random_dna}
\alias{random_dna}
\title{Random DNA string}
\usage{
random_dna(l)
}
\arguments{
\item{l}{length of desired DNA string}
}
\value{
DNA string
}
\description{
This function creates a random DNA string
}
\examples{
random_dna(int)

}
