% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_rna.R
\name{random_rna}
\alias{random_rna}
\title{Function that returns a random RNA string}
\usage{
random_rna(l)
}
\arguments{
\item{l}{}
}
\value{
A string of RNA with a random sequence of a given length
}
\description{
Function that returns a random RNA string
}
\examples{
random_rna(10)
}
